pyproject.toml
[tool.poetry]
name = "django-db-opt-sample"
version = "0.1.0"
description = ""
authors = ["catalunha <catalunha.mj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.0"
django = "^5.1"
djangorestframework = "^3.15.2"
django-debug-toolbar = "^4.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
